SQLAlchemy: Clean up and refactor SQLAlchemy polyfills
... to `cratedb_toolkit.util.sqlalchemy`.
