+++ cratedb_toolkit/cfr/systable.py
@@ -27,9 +27,9 @@ import polars as pl
 import sqlalchemy as sa
 from tqdm import tqdm
 
-from cratedb_toolkit.sqlalchemy.patch import patch_encoder
 from cratedb_toolkit.util import DatabaseAdapter
 from cratedb_toolkit.util.cli import error_logger
+from cratedb_toolkit.util.sqlalchemy import patch_encoder
 from cratedb_toolkit.wtf.core import InfoContainer
 
 logger = logging.getLogger(__name__)
