--- cratedb_toolkit/sqlalchemy/__init__.py
@@ -1,1 +1,0 @@
-from .polyfill import check_uniqueness_factory, polyfill_autoincrement, polyfill_refresh_after_dml, refresh_table
